init react and chakra ui

+++ backend/controller/product.controller.js
@@ -31,12 +31,15 @@ export const createProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(400).json({success: false, message: 'Invalid product ID'});
+    }
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message: 'Product deleted successfully'});
     } catch (error) {
         console.error(error);
-        res.status(500).json({success: false, message: 'Product not found'});
+        res.status(500).json({success: false, message: 'Server error'});
     }
 }
